Add tfe_agent_pool datasource
Agent pools do not have a full resource lifecycle yet, so an associated
resource here in the provider is not supported - yet.

This has the interesting side effect of needing to create an agent pool
out-of-band of the provider, temporarily.

This datasource is being added now, however, to be able to configure a
workspace to use an organization's single pool.