Temporary preview build for configuring TFC Agents on workspaces
This commit allows the work in b773cfb and 0b83f39 to be used in
building this provider with the necessary changes made to its
dependency, go-tfe.

With it, you may configure a workspace to use Terraform Cloud Agents
with an organization's single agent pool allotted to it at this time:

```
resource "tfe_organization" "foo" {
  name  = "foo-bar"
  email = "admin@company.com"
}

data "tfe_agent_pool" "bar" {
  name = "Default"
  organization = tfe_organization.foo.name
}

resource "tfe_workspace" "baz" {
  name                  = "cool-workspace"
  organization          = tfe_organization.foo.id
  execution_mode        = "agent"
  agent_pool_id         = data.tfe_agent_pool.bar.id
}
```

NOTE: An agent pool currently does not have a full API lifecycle, and
thus cannot yet be supported fully by the provider as a resource. This
means that the agent pool above must be created via UI or API OUT OF
BAND of the provider. The datasource is provided purely as a convenience
to supply the proper identifier to the workspace.